Allow users to filter all or some of the outbound objects in the metadata.  This has shown to result in a dramatic performance increase.  Using the FifoServer (new backend) and filtering all objects we go from 82 FPS to 172 FPS (Pass action).  Using WSGI we go from 22 FPS to 27 FPS (this is due to the poorer performance of the existing WSGI backend and JSON.net).  Users will want to filter all objects when training a point nav task and can filter down to one or two objects for object nav tasks.