@AlvaroHG 2020-09-17
This is great, improvement in clarity as well, left a few comments some minimal optimizations and deprecation of a similar unused system. LGTM!